[#71]Refactor navigation for Android and Desktop reuse with minimal KMP changes
- Add jvmMain source set in build.gradle.kts with Jetpack Compose deps for Android and Desktop
- Move RencfsNavigation to jvmMain as PlatformNavigation actual implementation
- Move RencfsComposeMainApp to jvmMain with splash screen and navigation logic
- Add expect PlatformNavigation to commonMain for future iOS/Web support
- Update MainActivity and Main.kt to call RencfsComposeMainApp from jvmMain
