+++ rencfsMultiplatform/build.gradle.kts
@@ -80,6 +80,19 @@ kotlin {
                 api(deps.sqldelight.driver.jvm)
             }
         }
+        // Define jvmMain source set
+        val jvmMain by creating {
+            dependsOn(commonMain.get())
+            dependencies {
+                api(compose.material3)
+                api(compose.material3AdaptiveNavigationSuite)
+                implementation(deps.jetbrains.androidx.navigation.compose)
+            }
+        }
+
+        // Link jvmMain to Android and Desktop
+        androidMain.dependsOn(jvmMain)
+        desktopMain.dependsOn(jvmMain)
     }
 }
